refactor: 페이지네이션 구현을 위한 공통 응답 DTO 생성
